@@ -3,13 +3,14 @@
   "version": "0.4.0",
   "private": true,
   "dependencies": {
-    "@angular/common": "5.2.7",
-    "@angular/compiler": "5.2.7",
-    "@angular/core": "5.2.7",
-    "@angular/platform-browser": "5.2.7",
-    "@angular/platform-browser-dynamic": "5.2.7",
-    "rxjs": "5.5.6",
-    "zone.js": "0.8.20"
+    "@angular/common": "6.1.10",
+    "@angular/compiler": "6.1.10",
+    "@angular/core": "6.1.10",
+    "@angular/platform-browser": "6.1.10",
+    "@angular/platform-browser-dynamic": "6.1.10",
+    "rxjs": "6.0.0",
+    "rxjs-compat": "6.0.0",
+    "zone.js": "0.8.29"
   },
   "devDependencies": {
     "awesome-typescript-loader": "3.4.1",
@@ -19,7 +20,7 @@
     "file-loader": "1.1.6",
     "html-webpack-plugin": "2.30.1",
     "style-loader": "0.19.1",
-    "typescript": "2.6.2",
+    "typescript": "2.7.2",
     "url-loader": "0.6.2",
     "webpack": "3.10.0",
     "webpack-dev-server": "2.9.7"
