@@ -3,7 +3,7 @@ import {HttpClient} from '@angular/common/http';
 
 import 'rxjs/add/operator/map';
 import {AppVersion} from "./model/appVersion";
-import {Observable} from "rxjs/Observable";
+import {Observable} from 'rxjs';
 
 @Injectable()
 export class AppVersionService {
